requirements.txt
discord
openai
python-dotenv
requests
pymysql